+++ README.md
@@ -24,14 +24,30 @@ yarn test
 yarn run coverage
 ```
 
-## Deployed contracts
+## Deployment
+
+1. Deploy the `Governance.sol`
+```
+npx hardhat run --network {NETWORK} ./scripts/deploy-governance.js
+```
+2. Initialise the Contract
+   1. Populate `GOV` and `KERNEL` properties in the `./init-governance.ts` script
+   2. Execute:
+```
+npx hardhat run --network {NETWORK} ./scripts/init-governance.js
+```
+3. Verify on Etherscan:
+```
+npx hardhat verify --network {NETWORK} {GOVERNANCE_ADDRESS}
+```
+
 ### Mainnet
 ```shell
 Governance deployed at:  TBD
 ```
 ### Rinkeby
 ```shell
-Governance deployed at:  TBD
+Governance deployed at:  0x820dB01Bf63f1CFCC0Cff60aE0a177d83678e633
 ```
 
 ## Audits
